Método jogarDado labirinto

+++ MatrizInimigos.java
@@ -3,7 +3,7 @@ public class MatrizInimigos{
   public static void main(String[] args){
     
     String[] inimigos = new String[3]; //vetor para os inimigos
-    double[][] dados = new double[3][2]; //matriz para os dados
+    int[][] dados = new int[3][2]; //matriz para os dados
     
     for(int i = 0; i < inimigos.length; i++){
       inimigos[i] = EntradaScanner.leiaString("Informe o nome do inimigo: ");
